+++ test/examples/nestedmodule1.js
@@ -2,6 +2,6 @@
 ({
   name: 'nestedmodule1',
   value: 1,
-  nested: require('./test/examples/nestedmodule2.js'),
+  nested: require('./nestedmodule2.js'),
 });
 /* eslint-enable */